improved memory usage in simulation; improved J-Matirx computation
Last layer neuron outputs are stored only once now.
The check whether to compute the jacobian matrix or not is now infront
of the loop, hence the loop gets entered only if necessary. The
sensitivities of the last layer can be computed all at once, which is a
major performance improvement.