refactored output value, improved performance
by using a hashmap to save the used indexes, it is faster to check if
one index is used already. The hash of the key can be calculated within
a fixed time, compared to search through all saved indexes in a vector,
which takes more time as the vector grows.